migrate_to_material_modals
Migrate Modal to TS
